refactor: update TaskDefinitionController, SchedulerController
